+++ lib/src/layouts/compact/yaru_navigation_rail_item.dart
