+++ lib/src/layouts/master_detail/yaru_detail_page.dart
