+++ lib/src/layouts/master_detail/yaru_landscape_layout.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'yaru_master_detail_page.dart';
@@ -164,15 +163,6 @@ class _YaruLandscapeLayoutState extends State<YaruLandscapeLayout> {
   }
 
   Widget _buildVerticalSeparator() {
-    // Fix for the ultra white divider on flutter web
-    final darkAndWeb =
-        kIsWeb && Theme.of(context).brightness == Brightness.dark;
-    if (darkAndWeb) {
-      return const VerticalDivider(
-        thickness: 0,
-        width: 0.01,
-      );
-    }
     return const VerticalDivider(
       thickness: 1,
       width: 1,
