+++ lib/yaru_widgets.dart
@@ -3,34 +3,35 @@ library yaru_widgets;
 // Constants
 export 'src/constants.dart';
 // Controls
+export 'src/controls/progress_indicator/yaru_circular_progress_indicator.dart';
+export 'src/controls/progress_indicator/yaru_linear_progress_indicator.dart';
+export 'src/controls/togglable/yaru_checkbox.dart';
+export 'src/controls/togglable/yaru_radio.dart';
+export 'src/controls/togglable/yaru_switch.dart';
+export 'src/controls/toggle_button/yaru_check_button.dart';
+export 'src/controls/toggle_button/yaru_radio_button.dart';
+export 'src/controls/toggle_button/yaru_switch_button.dart';
+export 'src/controls/toggle_button/yaru_toggle_button.dart';
+export 'src/controls/toggle_button/yaru_toggle_button_theme.dart';
 export 'src/controls/yaru_back_button.dart';
-export 'src/controls/yaru_check_button.dart';
-export 'src/controls/yaru_checkbox.dart';
 export 'src/controls/yaru_close_button.dart';
 export 'src/controls/yaru_color_disk.dart';
 export 'src/controls/yaru_icon_button.dart';
 export 'src/controls/yaru_option_button.dart';
 export 'src/controls/yaru_popup_menu_button.dart';
-export 'src/controls/yaru_progress_indicator.dart';
-export 'src/controls/yaru_radio.dart';
-export 'src/controls/yaru_radio_button.dart';
-export 'src/controls/yaru_switch.dart';
-export 'src/controls/yaru_switch_button.dart';
 export 'src/controls/yaru_title_bar.dart';
-export 'src/controls/yaru_toggle_button.dart';
-export 'src/controls/yaru_toggle_button_theme.dart';
 export 'src/controls/yaru_window_control.dart';
 // Extensions
 export 'src/extensions/border_radius_extension.dart';
 // Layouts
-export 'src/layouts/yaru_compact_layout.dart';
-export 'src/layouts/yaru_compact_layout_theme.dart';
-export 'src/layouts/yaru_detail_page.dart';
-export 'src/layouts/yaru_master_detail_page.dart';
-export 'src/layouts/yaru_master_detail_theme.dart';
-export 'src/layouts/yaru_master_tile.dart';
-export 'src/layouts/yaru_navigation_rail.dart';
-export 'src/layouts/yaru_navigation_rail_item.dart';
+export 'src/layouts/compact/yaru_compact_layout.dart';
+export 'src/layouts/compact/yaru_compact_layout_theme.dart';
+export 'src/layouts/compact/yaru_navigation_rail.dart';
+export 'src/layouts/compact/yaru_navigation_rail_item.dart';
+export 'src/layouts/master_detail/yaru_detail_page.dart';
+export 'src/layouts/master_detail/yaru_master_detail_page.dart';
+export 'src/layouts/master_detail/yaru_master_detail_theme.dart';
+export 'src/layouts/master_detail/yaru_master_tile.dart';
 // Pages
 export 'src/pages/yaru_section.dart';
 export 'src/pages/yaru_tabbed_page.dart';
